Remove recently introduced README symbolic link.
This patch deletes the symbolic link from README to README.md.  This
makes it easier for those of us using Windows and MSYS2 to build the
library.  You cannot create a symbolic link to a file that does not
exist in Windows, so when trying to extract the tar archive (using
MSYS2's tar utility), this is what happens:

    $ wget https://github.com/martinh/libconfuse/archive/master.tar.gz
    $ tar -xf master.tar.gz
    tar: libconfuse-master/README: Cannot create symlink to ‘README.md’: No such file or directory
    tar: Exiting with failure status due to previous errors

I am sure there are workarounds we could use, but it would be nice if we
didn't have to.  And I suppose that when you release a tarball for the
next version of the library, it probably won't have a symbolic link it,
so maybe it's not a big deal if the git repository has one.

Signed-off-by: David Grayson <davidegrayson@gmail.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>