Add release build target with md5sum generation + misc. cleanup
This patch simplifies release handling by adding a `make release` target
to the top level Makefile.  We distribute slightly different contents
for Windows (ZIP) on the one hand and all other systems (tarballs) on
the other, so it makes sense to simplify the release process for the
maintainer.

Also included in this patch is some minor cleanup, automake syntax
updates and reduced output noise.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>