Add cfg_setmulti and cfg_opt_setmulti.
Consider that you have a command line parser, and have parsed a line into
tokens. It turns out that the first token, the command, is a
'set-cfg-variable' type of command and that the other tokens describe
the new value, and possibly what variable to set.

I.e. you are given a command "set foo-list bar gazonk zilch"

and as a response to this command you want to set the "foo-list" option
to { "bar", "gazonk", "zilch" }. There is no current API to do this in a
simple maner, if you want to cover up for the possibility of syntax
errors during parsing and arbitrary number of command line arguments.

So, add cfg_setmulti (and cfg_opt_setmulti) to support this.

Signed-off-by: Peter Rosin <peda@lysator.liu.se>