Simplify lexer and parser API using fmemopen()
This patch is a refactor of the parser interface to the lexical
analyzer.  Prior to this patch the lexer had two entry points,
one for file based access and one for buffer based access.

The lexer is better suited for file operations and by using the
fmemopen() API (POSIX 2008) buffers can be accessed in the same
manner as a file.

The resulting code is less complex and is further simplified and
fixed (for leaks) in the following commit(s).

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>