lexer.l: Refactor cfg_lexer_include() to use same entry point as parser
This patch further simplifies the lexer, here the cfg_lexer_include()
function, which can recursively include and parse up to ten .conf
files.  With this patch the yypush/yypop API is used throughout, this
ensures not only returning to the previous parsing context properly,
but also closing of locally opened files to prevent memory leaks.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>