Fix #92: Add support for option annotation/comments
This patch adds support for option annotation.  It's a very simple
implementation, but it allows for one-liner comments to be both read
and written from/to a file both retrieved and set using the API.

A small unit test called annotate.c, with a annotate.conf, is also
included as an example.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>