Refactor option annotation slightly
This patch moves the assignment of the option comment to *after* the
option reset by the parser.  This is not optimal but was the easiest
way to implement it at the moment.

Also, when using the API to set a comment we must first verify that the
new comment could be strdup()'ed, otherwise we must return CFG_FAIL.

The API documentation has been updated accordingly.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>