Make CFG_PTR_CB not trigger a segfault in e.g. cfg_free().
CFG_PTR_CB makes use of the def.parsed field, but that field is only
duplicated for lists and functions. Since cfg_free_opt_array() frees
all fields unconditionally, this patch matches that and takes the safe
route by duplicating all fields, regardless of option flags and option
type.

The fields that are duplicated unconditionally should be empty if they
are not used, so there should be no memory thrown away by duplicating
stuff that is not needed.

Signed-off-by: Peter Rosin <peda@axentia.se>