Allow addressing options below multi-sections with cfg_getopt()
If you have a multi-section "sub" with titles, i.e. with
CFGF_MULTI | CFGF_TITLE set, you can now access options in this
manner:
	cfg_getopt(cfg, "sub=name|option");
where name is the title of an instance of the "sub" sub-section.

If a title contains a | or starts with a ' character (or both, like the
below example), the whole title has to be quoted with single quotes,
and the odd characters (backslash and single quote) escaped with a
backslash, like this:
	cfg_getopt(cfg, "sub='\\'a-very\\'silly|option\\\\title'|option");
I.e., the above example shows how to, from C code, quote
	'a-very'silly|option\title
Since C also uses the backslash as the escape character in strings,
backslashes need one extra round of escaping, which is why they come
in pairs in the above example.

For CFGF_MULTI sections without titles, you can instead access options
in this manner:
	cfg_getopt(cfg, "sub=2|option");
where 2 instead is the index of the instance.

Note: The index notation is /not/ available for options that have
titles, since that would introduce ambiguities for numeric titles.

Note: The (questionable) use of
	cfg_getopt(cfg, "sub|option");
to access "option" in the first instance of "sub" even if it is
CFGF_MULTI is preserved (with or without titles).

Signed-off-by: Peter Rosin <peda@axentia.se>