Support for CFGF_KEYVAL section flag, free-form runtime string opts
This patch adds support for free-form key=value string options, created
at runtime when parsing a file or buffer.

The CFGF_KEYVAL flag is a section option, when applied any option that
doesn't exist is created dynamically when parsing the file/buffer.  The
options are always created as type CFGT_STR and are primarly intended
for programs that need free-form environment variables and similar.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>