configure: Drop maintainer mode, unsafe and messes up distcleancheck
The AM_MAINTAINER_MODE macro disables rebuild rules for files like
configure.ac and lexer.l, which in many cases prevent users from being
able to do changes to these files.  It also prevents the distcleancheck,
called by distcheck, to properly detect that lexer.c should be part of
the distribution archive, even though it's not listed specifically.

There is a good description of AM_MAINTAINER_MODE in the manual, which
also mentions that even the originator of the macro has stopped using it
himself.  As a maintainer, I've never seen any use-case for it myself,
so I'm dropping it from libConfuse.

https://www.gnu.org/software/automake/manual/html_node/maintainer_002dmode.html

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>