only apply search path logic to relative pathnames
Adding any directory to the search path via cfg_add_searchpath breaks
lookup of absolute paths. So change the logic in cfg_searchpath() to
ignore the search path when the given filename is absolute, and merely
check that for existence.

This is technically an ABI change, but the current behaviour is quite
unusual and unexpected.