Fix isspace() arguments.
Per ANSI and POSIX, isspace() only accepts arguments that
are valid 'unsigned char' or EOF.

The code passed 'char' arguments, which on many platforms are
'signed char'. These are automatically promoted to integers, keeping
the sign, so negative values like -61 are passed to isspace(),
which is not allowed per the standards.

Casting to 'unsigned char' makes sure the isspace argument is in the
allowed range.

Fixes core dump e.g. in i3status.

Closes #180.